LayoutTests/fast/selectors/nth-last-child-matching-first-expected.txt

Test cases of :nth-last-child() matching only the last child.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ":nth-last-child(1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(+1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(+1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(-n+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(-n+1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(-n+1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(-2n+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(-2n+1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(-2n+1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(-3n+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(-3n+1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(-3n+1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(odd):nth-last-child(1):nth-last-child(2n+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(1):nth-last-child(2n+1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(1):nth-last-child(2n+1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(odd):nth-last-child(+1):nth-last-child(2n+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(+1):nth-last-child(2n+1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(+1):nth-last-child(2n+1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(odd):nth-last-child(-n+1):nth-last-child(2n+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(-n+1):nth-last-child(2n+1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(-n+1):nth-last-child(2n+1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(odd):nth-last-child(-2n+1):nth-last-child(2n+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(-2n+1):nth-last-child(2n+1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(-2n+1):nth-last-child(2n+1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(odd):nth-last-child(-3n+1):nth-last-child(2n+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(-3n+1):nth-last-child(2n+1)').length should be 1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(odd):nth-last-child(-3n+1):nth-last-child(2n+1)')[0].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(-n-1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(-n-1)').length should be 0. Threw exception Error: SyntaxError: DOM Exception 12
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing ":nth-last-child(n+1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n+1)').length should be 5. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n+1)')[0].id should be testcase1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n+1)')[1].id should be testcase2. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n+1)')[2].id should be testcase3. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n+1)')[3].id should be testcase4. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n+1)')[4].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

Testing ":nth-last-child(n-1)"
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n-1)').length should be 5. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n-1)')[0].id should be testcase1. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n-1)')[1].id should be testcase2. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n-1)')[2].id should be testcase3. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n-1)')[3].id should be testcase4. Threw exception Error: SyntaxError: DOM Exception 12
FAIL document.querySelectorAll('#test-root >> :nth-last-child(n-1)')[4].id should be testcase5. Threw exception Error: SyntaxError: DOM Exception 12
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).
FAIL getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor should be rgb(10, 100, 200). Was rgb(255, 0, 0).

PASS successfullyParsed is true

TEST COMPLETE
